PipelineScanner/requirements.txt
alembic==1.17.0
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
bandit==1.8.6
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dparse==0.6.4
fastapi==0.120.0
filelock==3.20.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
mdurl==0.1.2
narwhals==2.9.0
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
protobuf==6.33.0
psutil==7.1.2
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safety==3.6.2
safety-schemas==0.0.16
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
stevedore==5.5.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
